Cleanup designatort interface, use appropriate size type

+++ src/ansi-c/c_typecheck_initializer.cpp
@@ -408,9 +408,9 @@ void c_typecheck_baset::do_designated_initializer(
 
   // first phase: follow given designator
 
-  for(unsigned i=0; i<designator.size(); i++)
+  for(size_t i=0; i<designator.size(); i++)
   {
-    unsigned index=designator[i].index;
+    size_t index=designator[i].index;
     const typet &type=designator[i].type;
     
     assert(type.id()!=ID_symbol);
@@ -863,7 +863,7 @@ exprt c_typecheck_baset::do_initializer_list(
      to_array_type(type).size().is_nil())
   {
     // make complete by setting array size
-    unsigned size=result.operands().size();
+    size_t size=result.operands().size();
     result.type().id(ID_array);
     result.type().set(ID_size, from_integer(size, index_type()));
   }
